refactor: `CustomBottomNavigationBar` 파일명 변경
- `bottom_navigation_bar.dart` -> `custom_bottom_navigation_bar.dart` 파일명 변경
