+++ lib/screens/main_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:linring_front_flutter/screens/chat_screen.dart';
 import 'package:linring_front_flutter/screens/setting_screen.dart';
 import 'package:linring_front_flutter/screens/tag_show_screen.dart';
-import 'package:linring_front_flutter/widgets/bottom_navigation_bar.dart';
+import 'package:linring_front_flutter/widgets/custom_bottom_navigation_bar.dart';
 
 class MainScreen extends StatefulWidget {
   const MainScreen({Key? key}) : super(key: key);
